Created an add task page, chnages to the ui scheme.
Made some changes to the css file as well as some intext styling to
change the new theme seen currently.
Added the add task page and created a link to it from the view task page
which is currently the index.php
Added a finishedWork table to our database with exactly the same columns
as the task table. this is to be used for the finished work page which
shows the tasks that have been ended. meaning when the end task link is
clicked it is suppose to be moved to the finishedWork table and deleted
from the task table